Rework long telegramids

+++ Werewolf for Telegram/Werewolf Control/Models/CustomGifData.cs
@@ -12,7 +12,7 @@ namespace Werewolf_Control.Models
         public bool? Approved { get; set; } = null;
         public string DenyReason { get; set; }
         public bool NSFW { get; set; }
-        public int ApprovedBy { get; set; }
+        public long ApprovedBy { get; set; }
         public string VillagerDieImage { get; set; }
         public string WolfWin { get; set; }
         public string WolvesWin { get; set; }
